get rid of datasource
