+++ api/tests/test_importer.py
@@ -9,7 +9,7 @@ from app.all import import_all_from_env
 
 @mock.patch("app.plugins.collection", return_value=Mock())
 @mock.patch("app.config_repo.import_default_config_from_env", return_value=None)
-@mock.patch("app.datasource.import_datasources_from_env", return_value=None)
+@mock.patch("app.datasource_repo.import_datasources_from_env", return_value=None)
 @pytest.mark.asyncio
 async def test_import_plugins_from_env(mock_datasources, mock_plugins, mock_collection):
     os.environ[
@@ -31,7 +31,7 @@ async def test_import_plugins_from_env(mock_datasources, mock_plugins, mock_coll
 @mock.patch("app.config_repo.collection", return_value=Mock())
 @mock.patch("app.config_repo.get", return_value=None)
 @mock.patch("app.plugins.import_plugins_from_env", return_value=None)
-@mock.patch("app.datasource.import_datasources_from_env", return_value=None)
+@mock.patch("app.datasource_repo.import_datasources_from_env", return_value=None)
 @pytest.mark.asyncio
 async def test_import_feature_flags_from_env(
     mock_datasources, mock_plugins, mock_get, mock_collection
@@ -67,7 +67,7 @@ async def test_import_all_from_env_with_broken_plugin(mock_collection):
 @mock.patch("app.config_repo.collection", return_value=Mock())
 @mock.patch("app.config_repo.get")
 @mock.patch("app.plugins.import_plugins_from_env", return_value=None)
-@mock.patch("app.datasource.import_datasources_from_env", return_value=None)
+@mock.patch("app.datasource_repo.import_datasources_from_env", return_value=None)
 @pytest.mark.asyncio
 async def test_import_feature_flags_from_env_update(
     mock_datasource, mock_plugins, mock_get, mock_collection
@@ -93,7 +93,7 @@ async def test_import_feature_flags_from_env_update(
 @mock.patch("app.config_repo.collection", return_value=Mock())
 @mock.patch("app.config_repo.get")
 @mock.patch("app.plugins.import_plugins_from_env", return_value=None)
-@mock.patch("app.datasource.import_datasources_from_env", return_value=None)
+@mock.patch("app.datasource_repo.import_datasources_from_env", return_value=None)
 @pytest.mark.asyncio
 async def test_import_feature_flags_from_env_no_insert_or_update(
     mock_datasource, mock_plugins, mock_get, mock_collection
@@ -113,7 +113,7 @@ async def test_import_feature_flags_from_env_no_insert_or_update(
     mock_collection.return_value.insert_one.assert_not_called()
 
 
-@mock.patch("app.datasource.datasource_repo", return_value=Mock())
+@mock.patch("app.datasource_repo.import_datasources_from_env", return_value=Mock())
 @mock.patch("app.plugins.import_plugins_from_env", return_value=None)
 @mock.patch("app.config_repo.import_default_config_from_env", return_value=None)
 @pytest.mark.asyncio
